requirements.txt
openai
python-dotenv
faiss-cpu
numpy
gradio
